Cetic.11 Fix
The event Cetic.11 - An event which I think the Anniversary Patch added back in - Was broken.

When a new character siezes the Imperial Throne, this event fires 4 times, one for each of the holdings the character gets (California, and the Kingdom, Duchy and County of Sacramento)

I've added a character flag so it should only fire once per character.

Also - Since a character without  either Diplomacy or Learning 14+ can't hold onto California - I've made this a condition of taking the "Claim Imperial Throne" decision